Parse products from list
The products are now read from 'Streeplijst_####-##.csv' making it
possible to add new products by placing them in the csv file. If no
image is supplied with the same name as the product, a placeholder image
will be used.
Also, some keys are now bound to some standard products and will add
them to the shopping list. Shift+key will substract them from the list.